fix: remove dead and unused cli codes
