+++ api/pkg/cli/live/event_bus.go
@@ -6,21 +6,12 @@ type EventType string
 const (
 	EventAuth             EventType = "auth"
 	EventConfig           EventType = "config"
-	EventEnvCheck         EventType = "env_check"
-	EventFileDetected     EventType = "file_detected"
 	EventConnecting       EventType = "connecting"
 	EventConnected        EventType = "connected"
 	EventDisconnected     EventType = "disconnected"
 	EventReconnecting     EventType = "reconnecting"
 	EventSyncStart        EventType = "sync_start"
 	EventSyncProgress     EventType = "sync_progress"
-	EventSyncDone         EventType = "sync_done"
-	EventBuildStarted     EventType = "build_started"
-	EventBuildFailed      EventType = "build_failed"
-	EventBuildSuccess     EventType = "build_success"
-	EventDeploying        EventType = "deploying"
-	EventDeployed         EventType = "deployed"
-	EventDeployFailed     EventType = "deploy_failed"
 	EventError            EventType = "error"
 	EventInfo             EventType = "info"
 	EventPipelineProgress EventType = "pipeline_progress"
@@ -33,37 +24,6 @@ const (
 // EventPayload is the interface for type-safe event data.
 // Each event type that carries structured data gets its own struct.
 type EventPayload interface{ eventPayload() }
-
-// BuildFailedPayload carries build failure details.
-type BuildFailedPayload struct {
-	Logs  []string
-	Error string
-}
-
-func (BuildFailedPayload) eventPayload() {}
-
-// SyncProgressPayload carries file sync counters.
-type SyncProgressPayload struct {
-	FilesSynced int
-	Total       int
-}
-
-func (SyncProgressPayload) eventPayload() {}
-
-// EnvCheckPayload carries environment variable check results.
-type EnvCheckPayload struct {
-	Missing []string
-	Found   []string
-}
-
-func (EnvCheckPayload) eventPayload() {}
-
-// DeployedPayload carries deployment completion details.
-type DeployedPayload struct {
-	URL string
-}
-
-func (DeployedPayload) eventPayload() {}
 
 // PipelineProgressPayload carries real-time progress from the pipeline agent.
 type PipelineProgressPayload struct {
@@ -98,7 +58,7 @@ type DeploymentStatusPayload struct {
 
 func (DeploymentStatusPayload) eventPayload() {}
 
-// ApprovalNeededPayload carries the Dockerfile proposal for human approval.
+// ApprovalNeededPayload carries a proposal for human approval (Dockerfile, config, or other).
 type ApprovalNeededPayload struct {
 	Dockerfile      string
 	Summary         string
@@ -122,13 +82,7 @@ type Event struct {
 
 // isPriority returns true for events that must never be dropped.
 func (e Event) isPriority() bool {
-	if e.NeedsLLM ||
-		e.Type == EventError ||
-		e.Type == EventDeployed ||
-		e.Type == EventBuildFailed ||
-		e.Type == EventDeployFailed ||
-		e.Type == EventBuildSuccess ||
-		e.Type == EventApprovalNeeded {
+	if e.NeedsLLM || e.Type == EventError || e.Type == EventApprovalNeeded {
 		return true
 	}
 	// Build status errors are priority so the CLI always shows them
